Added more Brainfuck console functionality
Added the following commands:

1) 'HELPPLZ!!!'

Provides users information about the aliases used when
writing Brainfuck code against one of the troll interpreters.

2) 'ARRPLZ'

Allows users to see the current state of 'array' at
any location or range.

3) 'PTRPLZ'

Allows users to see the current index that 'ptr' is
pointing to in 'array'

Added a 'random' option for interpreter that allows
an interpreter to be chosen at random.

Added two new flags. The '-c' flag allows Brainfuck
code to be passed in as string and to be interpreted.
The '-v' flag displays the current version of the Brainfuck
interpeter/codebase.